add node.js file for safe execution in non-standard situations
